requirements.txt
flask
requests
openai
beautifulsoup4
python-dotenv
twikit
